Fix native memory safety, Liquify, SLAM boot-up, Vulkan, ink prediction; UX and model fixes
Executes the GLEE audit remediation plan (phases 1-8; phase 0 was verification only).

Native (core/nativebridge):
- Fix an integer-overflow-defeated bounds check and missing OpenCV type validation in
  nativeRestoreWallFeatureMap (GraffitiJNI.cpp) -- a malformed .gxr save could drive an
  out-of-bounds Mat read/write or a hard native crash.
- Remove gLastSplatTrace: a file-static in GraffitiJNI.cpp silently shadowed the real global in
  MobileGS.cpp, so the getter could never return anything; neither was ever written to or read by
  any caller.
- Add the same stride/RGBA_8888 checks the Vulkan stamp path already uses to Liquify's bitmap JNI
  boundary (nativePrepareLiquify/nativeBakeLiquify).
- Fix Liquify actually doing nothing: ImageWarper::initGl() had no caller anywhere in the app (the
  only GL context in the app is the unrelated 3D model viewer's GLSurfaceView), so every stroke's
  bake silently read back an unrendered texture and reported success. Add LiquifyGlContext, a
  dedicated headless EGL/pbuffer thread, and route every native Liquify call through it. Also
  guard bakeToBitmap on mInitialized, fix a 4-channel bitmap reaching the 3-channel-only
  LowLightEnhancer, and bound the per-touch-sample vertex scan to a padded region around the
  brush instead of scanning the whole (mGridDim+1)^2 mesh.
- Stop Liquify from booting the full SLAM engine (MobileGS: two permanent background threads, an
  ORB detector) for a tool that only ever touches the native ImageWarper -- add a
  nativeInitImageWarper() that constructs just that, and wire SlamManager.destroy() into
  EditorViewModel.onCleared() so the new GL thread/context also gets torn down.
- Vulkan stamp/stamp_masked shaders: query the device's actual maxComputeWorkGroupInvocations and
  fall back to an 8x8 workgroup variant when the fixed 16x16 (256 invocations) exceeds it -- the
  Vulkan spec's guaranteed minimum is only 128. Floor each dab's radius in the C++ dispatch bbox to
  match the shader's own floor.
- InkStrokePredictorJNI: stop truncating the predicted timestamp through jfloat (silently wrong
  past ~4.6h of device uptime -- widen to jdouble end to end), and correct four Kalman/wobble
  tuning constants that deviated ~6x from the file's own documented derivation formula.
- Delete a fully duplicated nested copy of the vendored GLM library (libs/glm/glm/glm, 3.2MB/429
  files, unreferenced) and the dead, duplicated NativeUtil.h.

UX (app):
- Wire EditorViewModel.railExpansion (built and persisted, never connected to any UI) into every
  rail host item's initiallyExpanded/onExpandedChange in MainActivity.kt's ConfigureRailItems.
- Fix ShortcutsSheet's active tab re-seeding itself (and visibly snapping) the instant a long-press
  on the sheet itself changed the favorites count.
- Replace StoreChooserDialog's AlertDialog with FloatingWindow, matching every other panel in the
  app's non-modal chrome contract (ConfirmDialog's own AlertDialog stays -- it deliberately gates a
  destructive action).

Model/conceptual (core/common, feature/editor):
- Remove LayerType.FILTER/MASK (constructed nowhere in the codebase) and actually set
  type = LayerType.VECTOR at the two shape-layer construction sites that used to leave it
  defaulted to RASTER.
- Give Animation Assist a real non-frame top-level layer: Layer.isPinnedAcrossFrames, excluded
  from AnimationFrames.topLevelFrames/frame-index math and always rendered at full opacity, with a
  per-layer toggle in the layer's hidden menu.
- Update ARCHITECTURE.md's module table: it was missing :core:engine and :desktop, which changes
  the shared-module coupling story from two consumers (Graffux, GraffitiXR) to three.

Not executed (see PR description): a full rail-as-IA navigation redesign and removing azphalt's
sandboxed code execution in favor of a declarative filter/LUT DSL -- both large, hard-to-verify-
without-a-device architectural changes flagged for the repo owner's own call, not made unilaterally.

+++ core/nativebridge/src/main/java/com/hereliesaz/graffitixr/nativebridge/InkStrokePredictor.kt
@@ -49,11 +49,13 @@ class InkStrokePredictor : AutoCloseable {
         if (nativeHandle == 0L || !hasInput) return null
         val values = nativePredict(nativeHandle) ?: return null
         if (values.size < 4) return null
+        // A double round-trips uptimeMillis exactly (a jfloat can't past ~4.6h of device uptime --
+        // see nativePredict's doc comment); position/pressure just widen back down from it losslessly.
         return Prediction(
-            x = values[0],
-            y = values[1],
+            x = values[0].toFloat(),
+            y = values[1].toFloat(),
             uptimeMillis = values[2].toLong(),
-            pressure = values[3].takeIf { it.isFinite() && it >= 0f } ?: 1f,
+            pressure = values[3].toFloat().takeIf { it.isFinite() && it >= 0f } ?: 1f,
         )
     }
 
@@ -75,6 +77,6 @@ class InkStrokePredictor : AutoCloseable {
         pressure: Float,
         isDown: Boolean,
     ): Boolean
-    private external fun nativePredict(handle: Long): FloatArray?
+    private external fun nativePredict(handle: Long): DoubleArray?
     private external fun nativeDestroy(handle: Long)
 }
